Create problemIV.java
Hope you got used to menu driven programming. In this exercise you need to perform updations.
For instance, lets suppose the user is asking for an email update. You need to give provision for updating the email address.
 
Lets make the problem simple.
These are the following tasks to be done:
First register the guest/customer,
Ask for update in email,
Update the email if asked for, and
Display the details.